requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
loguru==0.7.2
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
urllib3==2.2.3
